app/garden_db_project/horticulture/weather_api.py
"""
Weather API integration module for the Gardening Database.

This module provides integration with the Open Weather API for supplemental environmental context.
As per the PRD, usage is limited to 500 calls per 24 hours.

Note: This is a placeholder module for future implementation.
"""

import requests
import logging
from django.conf import settings
from datetime import datetime, timedelta

logger = logging.getLogger(__name__)

# Constants
API_CALL_LIMIT = 500  # Maximum calls per 24 hours as per PRD
API_BASE_URL = "https://api.openweathermap.org/data/2.5"

class WeatherAPIClient:
    """
    Client for interacting with the Open Weather API.
    
    This is a placeholder implementation that will be expanded in the future.
    """
    
    def __init__(self):
        """Initialize the Weather API client."""
        # API key would be stored in settings.py in a real implementation
        self.api_key = getattr(settings, 'OPEN_WEATHER_API_KEY', None)
        self.call_counter = 0
        self.reset_date = datetime.now() + timedelta(days=1)
    
    def _check_rate_limit(self):
        """
        Check if we've exceeded our rate limit.
        
        Returns:
            bool: True if we can make another call, False otherwise.
        """
        # Reset counter if a new day has started
        if datetime.now() > self.reset_date:
            self.call_counter = 0
            self.reset_date = datetime.now() + timedelta(days=1)
        
        # Check if we've hit the limit
        if self.call_counter >= API_CALL_LIMIT:
            logger.warning("Weather API call limit reached (500 calls/24h)")
            return False
        
        return True
    
    def get_current_weather(self, latitude, longitude):
        """
        Get current weather for a location.
        
        Args:
            latitude (float): Latitude coordinate
            longitude (float): Longitude coordinate
            
        Returns:
            dict: Weather data or None if API call failed
        """
        if not self.api_key:
            logger.error("No Weather API key configured")
            return None
            
        if not self._check_rate_limit():
            return None
            
        # This would make an actual API call in a real implementation
        # url = f"{API_BASE_URL}/weather?lat={latitude}&lon={longitude}&appid={self.api_key}&units=imperial"
        # response = requests.get(url)
        # self.call_counter += 1
        
        # Placeholder return
        logger.info(f"Would fetch weather for coordinates: {latitude}, {longitude}")
        return {
            "location": f"{latitude}, {longitude}",
            "temperature": "N/A",
            "humidity": "N/A",
            "conditions": "Placeholder data - API not implemented yet"
        }
    
    def get_forecast(self, latitude, longitude, days=5):
        """
        Get weather forecast for a location.
        
        Args:
            latitude (float): Latitude coordinate
            longitude (float): Longitude coordinate
            days (int): Number of days to forecast
            
        Returns:
            dict: Forecast data or None if API call failed
        """
        if not self.api_key:
            logger.error("No Weather API key configured")
            return None
            
        if not self._check_rate_limit():
            return None
            
        # This would make an actual API call in a real implementation
        # url = f"{API_BASE_URL}/forecast?lat={latitude}&lon={longitude}&appid={self.api_key}&units=imperial"
        # response = requests.get(url)
        # self.call_counter += 1
        
        # Placeholder return
        logger.info(f"Would fetch {days}-day forecast for coordinates: {latitude}, {longitude}")
        return {
            "location": f"{latitude}, {longitude}",
            "forecast": [
                {
                    "day": f"Day {i+1}",
                    "temperature": "N/A",
                    "conditions": "Placeholder data - API not implemented yet"
                }
                for i in range(days)
            ]
        }

# Singleton instance for use throughout the application
weather_client = WeatherAPIClient()
