app/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.0
graphene-django==3.1.5
gunicorn==21.2.0
whitenoise==6.5.0
Pillow==10.1.0
django-environ==0.11.2
django-extensions==3.2.3
django-crispy-forms==2.0
crispy-bootstrap5==0.7
drf-yasg==1.21.7
django-debug-toolbar==4.2.0
